# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Betayin Kino",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/comment/add": {
            "post": {
                "description": "Adds a new comment with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Add a new comment",
                "parameters": [
                    {
                        "description": "Comment Add Request",
                        "name": "createCommentReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/movieservice.CreateCommentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/movieservice.CreateCommentRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/comment/all": {
            "get": {
                "description": "Retrieves a list of all comments with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Get all comments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by movie ID",
                        "name": "movie_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by rating",
                        "name": "rate",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of comments",
                        "schema": {
                            "$ref": "#/definitions/movieservice.GetAllCommentsRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/comment/delete": {
            "delete": {
                "description": "Deletes a comment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Delete a comment",
                "parameters": [
                    {
                        "description": "Comment Delete Request",
                        "name": "deleteCommentReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/movieservice.DeleteCommentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/movieservice.DeleteCommentRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/comment/update": {
            "put": {
                "description": "Updates the details of a comment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Update comment details",
                "parameters": [
                    {
                        "description": "Comment Update Request",
                        "name": "updateCommentReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/movieservice.UpdateCommentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/movieservice.UpdateCommentRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/comment/{id}": {
            "get": {
                "description": "Retrieves details of a comment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Get comment details by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/movieservice.GetCommentRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "movieservice.CommentModel": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "movie_id": {
                    "type": "string"
                },
                "rate": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "movieservice.CreateCommentReq": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "movie_id": {
                    "type": "string"
                },
                "rate": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "movieservice.CreateCommentRes": {
            "type": "object",
            "properties": {
                "comment": {
                    "$ref": "#/definitions/movieservice.CommentModel"
                }
            }
        },
        "movieservice.DeleteCommentReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "movieservice.DeleteCommentRes": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "movieservice.GetAllCommentsRes": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/movieservice.CommentModel"
                    }
                }
            }
        },
...
        "movieservice.GetCommentRes": {
            "type": "object",
            "properties": {
                "comment": {
                    "$ref": "#/definitions/movieservice.CommentModel"
                }
            }
        },
...
        "movieservice.UpdateCommentReq": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "rate": {
                    "type": "integer"
                }
            }
        },
        "movieservice.UpdateCommentRes": {
            "type": "object",
            "properties": {
                "comment": {
                    "$ref": "#/definitions/movieservice.CommentModel"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authourization",
            "in": "header"
        }
    }
}